Adapted the Arm DMP to new data
The Controller has been changed to the new Map format and has been
tested with the new controller. For the moment the arm is moving but
with jumps. I think the problem is in the trajectory reaching some
points not reachable with the EF controll. Future work change the EF for
Velocities.