Added Two method For the Trajectory Learning
Find a Problem in the real world experiments. Where a change in the
position of the world ( most of times the orientation ) make the robot
end a in a different position. In this methods the robot use itself as a
centre of the frame. Some strange problems computing position and
orientations there are some works around which will be fixed latter